DerivedPlaceholderModelクラスを追加し、DynamicTemplateConversionServiceImplでの派生プレースホルダーの処理を改善
